Export transaction labelling onto Store class to also use in transaction modal
